llk -> arch: refactor sanitizer assertions
